requirements.txt
einops
hydra-core
lightning
lion_pytorch
matplotlib
numpy
omegaconf
pandas
pytorch_lightning
Requests
scikit_learn
scipy
seaborn
torch
torch_ema
transformers
wandb
pykan